requirements.txt
boto3~=1.26.67
PyYAML~=6.0
botocore~=1.29.67